# perf(ext/node): Optimise Buffer string operations